+++ CHANGELOG.md
@@ -81,6 +81,7 @@
 * Handle non-404 4xx errors in UserDetail. Refs UIU-3057.
 * Update the `index.all` translation. Refs UIU-1027.
 * Select user roles' modal's "Select all" button selects all *visible* records, following convention. Refs UIU-3548.
+* Update IfPerm in user edit form from `ui-users-roles.view` to `ui-authorization-roles.users.settings.view` since `ui-users-roles.view` includes extraneous subpermissions. Refs UIU-3521.
 
 ## [12.1.16] (https://github.com/folio-org/ui-users/tree/v12.1.16) (2025-12-29)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v12.1.15...v12.1.16)